+++ packages/docusaurus-plugin-content-docs/src/client/index.ts
@@ -20,6 +20,41 @@ import {
 } from './docsClientUtils';
 import type {UseDataOptions} from '@docusaurus/types';
 
+export {
+  isDocsPluginEnabled,
+  useDocById,
+  findSidebarCategory,
+  findFirstSidebarItemLink,
+  isActiveSidebarItem,
+  isVisibleSidebarItem,
+  useVisibleSidebarItems,
+  useSidebarBreadcrumbs,
+  useDocsVersionCandidates,
+  useLayoutDoc,
+  useLayoutDocsSidebar,
+  useDocRootMetadata,
+  useCurrentSidebarCategory,
+  filterDocCardListItems,
+} from './docsUtils';
+
+export {useDocsPreferredVersion} from './docsPreferredVersion';
+
+export {
+  DocSidebarItemsExpandedStateProvider,
+  useDocSidebarItemsExpandedState,
+} from './docSidebarItemsExpandedState';
+
+export {DocsVersionProvider, useDocsVersion} from './docsVersion';
+
+export {DocsSidebarProvider, useDocsSidebar} from './docsSidebar';
+
+export {DocProvider, useDoc, type DocContextValue} from './doc';
+
+export {
+  useDocsPreferredVersionByPluginId,
+  DocsPreferredVersionContextProvider,
+} from './docsPreferredVersion';
+
 export type ActivePlugin = {
   pluginId: string;
   pluginData: GlobalPluginData;
